refactor(rpc): Move verifier::deposit_finalize ouside of rpc module.
