+++ core/src/rpc/error.rs
@@ -13,7 +13,7 @@ pub(crate) fn input_ended_prematurely() -> Status {
     Status::invalid_argument("Input stream ended prematurely")
 }
 
-pub(crate) fn sighash_stream_ended_prematurely() -> Status {
+pub(crate) fn _sighash_stream_ended_prematurely() -> Status {
     Status::internal("Sighash stream ended prematurely")
 }
 
@@ -21,7 +21,7 @@ pub(crate) fn output_stream_ended_prematurely() -> Status {
     Status::internal("Output stream ended prematurely".to_string())
 }
 
-pub(crate) fn sighash_stream_failed(msg: Status) -> Status {
+pub(crate) fn _sighash_stream_failed(msg: Status) -> Status {
     Status::internal(format!("Sighash stream failed: {msg}"))
 }
